python-cassandra/requirements.txt
flask
cassandra-driver